Moving string distance algorithms to a 'string' subfolder in rust code.
